chore: address 4/5 neal's comments
